+++ src/lasers.rs
@@ -2,7 +2,8 @@ use std::cmp::{max, min};
 use crate::huffman_code::HuffTree;
 use gpio::GpioValue::{High, Low};
 use gpio::{GpioIn, GpioOut};
-use std::thread;
+use std::{fs, thread};
+use std::collections::HashMap;
 use std::time::Duration;
 
 const LASER_PIN: u16 = 18;
@@ -209,4 +210,39 @@ impl Receiver {
             1.0 - error,
         );
     }
+}
+
+pub fn do_lasers() {
+    let mut laser = Laser::new();
+    let mut receiver = Receiver::new();
+    let mut message = fs::read_to_string("./src/lasers.rs").expect("error opening file");
+    // let mut message = "Hello World.".to_string();
+
+    // Compress message with Huffman Coding.
+    let mut freq_map = HashMap::new();
+    let mut huff_tree = HuffTree::new();
+    for char in message.chars() {
+        let cout = freq_map.entry(char).or_insert(0);
+        *cout += 1;
+    }
+    huff_tree.build_tree(freq_map);
+    let mut encoded_message = huff_tree.encode_string(&mut message);
+
+    // Start a thread each for the laser and receiver.
+    let receiver_thread = thread::Builder::new()
+        .name("receiver".to_string())
+        .spawn(move || loop {
+            receiver.print_message(&mut huff_tree);
+        });
+
+    let laser_thread = thread::Builder::new()
+        .name("laser".to_string())
+        .spawn(move || loop {
+            laser.send_message(encoded_message.as_mut());
+            thread::sleep(Duration::from_millis(2500))
+        });
+
+    laser_thread.unwrap().join().unwrap();
+    receiver_thread.unwrap().join().unwrap();
+
 }
